Update wsgiapp.handle_ValueError(...) to match Falcon 2.0+ interface.
The use of old signature is detected by Falcon, and automatically reordered
using a compatibility shim. Also, a warning is raised when running the app or
its test suite:

  src/mailman/rest/wsgiapp.py:202: DeprecatedWarning: handler is using a deprecated signature; please order its arguments as handle_ValueError(req, resp, ex, params). This compatibility shim will be removed in Falcon 5.0.

The ordering of the parameters passed to custom error handlers was adjusted in
Falcon 2.0:

  # Before
  def handle_error(ex, req, resp, params):
      pass

  # Falcon 2.0
  def handle_error(req, resp, ex, params):
      pass

See also: https://falcon.readthedocs.io/en/stable/changes/2.0.0.html.