docs: Use index.rst and create a new README file
Existing README.rst seems to fail rendering on PyPI and because of
that, PyPI refuses the upload of the package. In order to actually
resolve this, use the standard "index.rst" for the documentation
home page and create a new README.md which can be rendered on Gitlab
and PyPI along with the package which includes a simple link to
PyPI and Gitlab to report issues.

* docs: Use index.rst and create a new README file

Existing README.rst seems to fail rendering on PyPI and because of
that, PyPI refuses the upload of the package. In order to actually
resolve this, use the standard "index.rst" for the documentation
home page and create a new README.md which can be rendered on Gitlab
and PyPI along with the package which includes a simple link to
PyPI and Gitlab to report issues.