Add a mime_delete test for pass_extensions actually passes listed extension.
As a result of #1204, I thought there was a bug in processing mass_extensions and filter_extensions which would require them being entered with the leading dot, but I was mistaken. There is no bug, but I created this test for it, so we may as well keep the test.

* Fix qa issue.

* Add a mime_delete test for pass_extensions actually passes listed extension.